+++ clients/algoliasearch-client-scala/src/main/scala/algoliasearch/ingestion/SourceUpdateInput.scala
@@ -27,10 +27,10 @@ object SourceUpdateInputSerializer extends Serializer[SourceUpdateInput] {
 
     case (TypeInfo(clazz, _), json) if clazz == classOf[SourceUpdateInput] =>
       json match {
+        case value: JObject if value.obj.exists(_._1 == "projectID") => Extraction.extract[SourceBigQuery](value)
         case value: JObject => Extraction.extract[SourceUpdateCommercetools](value)
         case value: JObject => Extraction.extract[SourceJSON](value)
         case value: JObject => Extraction.extract[SourceCSV](value)
-        case value: JObject => Extraction.extract[SourceBigQuery](value)
         case value: JObject => Extraction.extract[SourceUpdateDocker](value)
         case _              => throw new MappingException("Can't convert " + json + " to SourceUpdateInput")
       }
@@ -38,10 +38,10 @@ object SourceUpdateInputSerializer extends Serializer[SourceUpdateInput] {
 
   override def serialize(implicit format: Formats): PartialFunction[Any, JValue] = { case value: SourceUpdateInput =>
     value match {
+      case value: SourceBigQuery            => Extraction.decompose(value)(format - this)
       case value: SourceUpdateCommercetools => Extraction.decompose(value)(format - this)
       case value: SourceJSON                => Extraction.decompose(value)(format - this)
       case value: SourceCSV                 => Extraction.decompose(value)(format - this)
-      case value: SourceBigQuery            => Extraction.decompose(value)(format - this)
       case value: SourceUpdateDocker        => Extraction.decompose(value)(format - this)
     }
   }