+++ snippets/scala/src/main/scala/Ingestion.scala
@@ -268,7 +268,7 @@ class SnippetIngestionClient {
 
   /** Snippet for the enableTask method.
     *
-    * enableTask
+    * enable task e2e
     */
   def snippetForIngestionClientEnableTask(): Unit = {
     // Initialize the client
@@ -276,7 +276,7 @@ class SnippetIngestionClient {
 
     // Call the API
     val res = client.enableTask(
-      taskID = "6c02aeb1-775e-418e-870b-1faccd4b2c0f"
+      taskID = "76ab4c2a-ce17-496f-b7a6-506dc59ee498"
     )
 
     // Use the response
@@ -444,7 +444,7 @@ class SnippetIngestionClient {
 
     // Call the API
     val res = client.getSource(
-      sourceID = "6c02aeb1-775e-418e-870b-1faccd4b2c0f"
+      sourceID = "75eeb306-51d3-4e5e-a279-3c92bd8893ac"
     )
 
     // Use the response
@@ -585,7 +585,11 @@ class SnippetIngestionClient {
     // Call the API
     val res = client.searchTasks(
       taskSearch = TaskSearch(
-        taskIDs = Seq("6c02aeb1-775e-418e-870b-1faccd4b2c0f", "947ac9c4-7e58-4c87-b1e7-14a68e99699a")
+        taskIDs = Seq(
+          "6c02aeb1-775e-418e-870b-1faccd4b2c0f",
+          "947ac9c4-7e58-4c87-b1e7-14a68e99699a",
+          "76ab4c2a-ce17-496f-b7a6-506dc59ee498"
+        )
       )
     )