+++ clients/algoliasearch-client-scala/CHANGELOG.md
@@ -1,3 +1,8 @@
+## [2.0.0-alpha.8](https://github.com/algolia/algoliasearch-client-scala/compare/2.0.0-alpha.7...2.0.0-alpha.8)
+
+- [6624759e1](https://github.com/algolia/api-clients-automation/commit/6624759e1) feat(specs): add CustomFields to CT Source Input ([#2742](https://github.com/algolia/api-clients-automation/pull/2742)) by [@damcou](https://github.com/damcou/)
+- [ef37e65b7](https://github.com/algolia/api-clients-automation/commit/ef37e65b7) fix(clients): add e2e tests for analytics API ([#2737](https://github.com/algolia/api-clients-automation/pull/2737)) by [@millotp](https://github.com/millotp/)
+
 ## [2.0.0-alpha.7](https://github.com/algolia/algoliasearch-client-scala/compare/2.0.0-alpha.6...2.0.0-alpha.7)
 
 - [b3d61128a](https://github.com/algolia/api-clients-automation/commit/b3d61128a) docs(clients): add ACL section to code doc ([#2576](https://github.com/algolia/api-clients-automation/pull/2576)) by [@millotp](https://github.com/millotp/)